No longer deletes index.php

+++ clear_cache.php
@@ -4,7 +4,8 @@ $dir = dirname(__FILE__) . '/cache/*';
 
 // Open a known directory, and proceed to read its contents  
 foreach(glob($dir) as $file)   {  
-  unlink($file);
+  if( !preg_match('/index\.php/', $file) )
+    unlink($file);
 }
 
 header('Location: /wp-admin/admin.php?page=wufoo-phooey');
